requirements.txt
# Core
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
pyyaml>=6.0.1
anthropic>=0.40.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
